+++ modules/groovy/exec/src/sap/commerce/toolset/groovy/exec/GroovyExecConstants.kt
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.groovy
+package sap.commerce.toolset.groovy.exec
 
 import com.intellij.icons.AllIcons
 import sap.commerce.toolset.groovy.exec.context.ReplicaSelectionMode
@@ -53,4 +53,9 @@ object GroovyExecConstants {
     }
 
     const val DEFAULT_WEB_CONTEXT = "default"
+
+    // See: de.hybris.platform.hac.scripting.impl.DefaultScriptingLanguageExecutor
+    const val RESPONSE_OUTPUT_TEXT = "outputText"
+    const val RESPONSE_EXECUTION_RESULT = "executionResult"
+    const val RESPONSE_STACKTRACE_TEXT = "stacktraceText"
 }