+++ modules/groovy/exec/src/sap/commerce/toolset/groovy/exec/context/GroovyReplicaAwareContext.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,7 +19,7 @@
 package sap.commerce.toolset.groovy.exec.context
 
 import sap.commerce.toolset.exec.context.ReplicaContext
-import sap.commerce.toolset.groovy.GroovyExecConstants
+import sap.commerce.toolset.groovy.exec.GroovyExecConstants
 
 data class GroovyReplicaAwareContext(
     val replicaSelectionMode: ReplicaSelectionMode,