+++ modules/groovy/ui/src/sap/commerce/toolset/groovy/actionSystem/GroovyExecuteAction.kt
@@ -33,7 +33,7 @@ import sap.commerce.toolset.groovy.editor.groovyExecContextSettings
 import sap.commerce.toolset.groovy.editor.groovySplitEditor
 import sap.commerce.toolset.groovy.exec.GroovyExecClient
 import sap.commerce.toolset.groovy.exec.context.GroovyExecContext
-import sap.commerce.toolset.groovy.groovyTransactionMode
+import sap.commerce.toolset.groovy.exec.groovyTransactionMode
 import sap.commerce.toolset.hac.actionSystem.ExecuteStatementAction
 import sap.commerce.toolset.hac.exec.HacExecConnectionService
 import sap.commerce.toolset.settings.state.TransactionMode